utils.py: add PipelineFileMatcher, use this in get_file_metadata
I think I may have overdesigned this a little bit with the `FileMatcher`
abstract base class, but the other designs that occurred to me seemed
even more clunky.

I kind of like asking an object whether something should be indexed, and
for relevant metadata if so, and it made sense to make a dummy version of
the real file matcher that allows everything (if there's no file manifest
included with the pipeline). Using the abstract base class as the type
annotation seems like a good idea.