Flip arity of `find_pipeline_manifests` and `get_absolute_workflows`
`get_absolute_workflows` is intended for use at the beginning of a DAG, for
constructing a list of absolute Paths to workflows on disk. The input to this
function is always a "hand-crafted" sequence of workflow Paths, so requiring
users of this function to build a List[Path] is just extra noise without any
benefit.

Use this modified function to adjust each DAG to keep a list of workflow CWL
files, instead of distinct variables `cwl_workflowN` for N ∊ ℕ.

With every DAG storing a List[Path] of absolute CWL files now, all usages of
`find_pipeline_manifests` would require this list to be unpacked, so adjust
that function to accept the list of CWL workflows directly.