Correction:
`if is_public is not None:` should be `if is_public is not True`

This is a relic from an earlier version where is_public was initialized as
None rather than false. While stitching together revisions into the
newer version of this script, this was mistakenly left unchanged.

This has been fixed.